add: links to platforms
